Add dry run checkbox in order to skip Git operations for debugging purposes

+++ Vss2Git/GitExporter.cs
@@ -40,6 +40,7 @@ namespace Hpdi.Vss2Git
         private readonly StreamCopier streamCopier = new StreamCopier();
         private readonly HashSet<string> tagsUsed = new HashSet<string>();
         private bool ignoreErrors = false;
+        private bool dryRun = false;
 
         private string emailDomain = "localhost";
         public string EmailDomain
@@ -68,6 +69,12 @@ namespace Hpdi.Vss2Git
             set { ignoreErrors = value; }
         }
 
+        public bool DryRun
+        {
+            get { return dryRun; }
+            set { dryRun = value; }
+        }
+
         public GitExporter(WorkQueue workQueue, Logger logger,
             RevisionAnalyzer revisionAnalyzer, ChangesetBuilder changesetBuilder)
             : base(workQueue, logger)
@@ -87,38 +94,44 @@ namespace Hpdi.Vss2Git
                 LogStatus(work, "Initializing Git repository");
 
                 // create repository directory if it does not exist
-                if (!Directory.Exists(repoPath))
+                if (!dryRun && !Directory.Exists(repoPath))
                 {
                     Directory.CreateDirectory(repoPath);
                 }
 
-                var git = new GitWrapper(repoPath, logger);
-                git.CommitEncoding = commitEncoding;
+                GitWrapper git = null;
 
-                while (!git.FindExecutable())
+                if (!dryRun)
                 {
-                    var button = MessageBox.Show("Git not found in PATH. " +
-                        "If you need to modify your PATH variable, please " +
-                        "restart the program for the changes to take effect.",
-                        "Error", MessageBoxButtons.RetryCancel, MessageBoxIcon.Error);
-                    if (button == DialogResult.Cancel)
+                    git = new GitWrapper(repoPath, logger);
+
+                    git.CommitEncoding = commitEncoding;
+
+                    while (!git.FindExecutable())
                     {
-                        workQueue.Abort();
+                        var button = MessageBox.Show("Git not found in PATH. " +
+                            "If you need to modify your PATH variable, please " +
+                            "restart the program for the changes to take effect.",
+                            "Error", MessageBoxButtons.RetryCancel, MessageBoxIcon.Error);
+                        if (button == DialogResult.Cancel)
+                        {
+                            workQueue.Abort();
+                            return;
+                        }
+                    }
+
+                    if (!RetryCancel(delegate { git.Init(); }))
+                    {
                         return;
                     }
-                }
 
-                if (!RetryCancel(delegate { git.Init(); }))
-                {
-                    return;
-                }
-
-                if (commitEncoding.WebName != "utf-8")
-                {
-                    AbortRetryIgnore(delegate
+                    if (commitEncoding.WebName != "utf-8")
                     {
-                        git.SetConfig("i18n.commitencoding", commitEncoding.WebName);
-                    });
+                        AbortRetryIgnore(delegate
+                        {
+                            git.SetConfig("i18n.commitencoding", commitEncoding.WebName);
+                        });
+                    }
                 }
 
                 var pathMapper = new VssPathMapper();
@@ -164,7 +177,7 @@ namespace Hpdi.Vss2Git
                     }
 
                     // commit changes
-                    if (needCommit)
+                    if (!dryRun && needCommit)
                     {
                         LogStatus(work, "Committing " + changesetDesc);
                         if (CommitChangeset(git, changeset))
@@ -212,14 +225,17 @@ namespace Hpdi.Vss2Git
                                     tagComment = labelName;
                                 }
 
-                                if (AbortRetryIgnore(
-                                    delegate
-                                    {
-                                        git.Tag(tagName, label.User, GetEmail(label.User),
-                                            tagComment, label.DateTime);
-                                    }))
+                                if (!dryRun)
                                 {
-                                    ++tagCount;
+                                    if (AbortRetryIgnore(
+                                        delegate
+                                        {
+                                            git.Tag(tagName, label.User, GetEmail(label.User),
+                                                tagComment, label.DateTime);
+                                        }))
+                                    {
+                                        ++tagCount;
+                                    }
                                 }
                             }
                         }
@@ -233,7 +249,10 @@ namespace Hpdi.Vss2Git
                 logger.WriteSectionSeparator();
                 logger.WriteLine("Git export complete in {0:HH:mm:ss}", new DateTime(stopwatch.ElapsedTicks));
                 logger.WriteLine("Replay time: {0:HH:mm:ss}", new DateTime(replayStopwatch.ElapsedTicks));
-                logger.WriteLine("Git time: {0:HH:mm:ss}", new DateTime(git.ElapsedTime.Ticks));
+                if (!dryRun)
+                {
+                    logger.WriteLine("Git time: {0:HH:mm:ss}", new DateTime(git.ElapsedTime.Ticks));
+                }
                 logger.WriteLine("Git commits: {0}", commitCount);
                 logger.WriteLine("Git tags: {0}", tagCount);
             });
@@ -345,8 +364,11 @@ namespace Hpdi.Vss2Git
                                     {
                                         if (((VssProjectInfo)itemInfo).ContainsFiles())
                                         {
-                                            git.Remove(targetPath, true);
-                                            needCommit = true;
+                                            if (!dryRun)
+                                            {
+                                                git.Remove(targetPath, true);
+                                                needCommit = true;
+                                            }
                                         }
                                         else
                                         {
@@ -393,8 +415,11 @@ namespace Hpdi.Vss2Git
                                     var projectInfo = itemInfo as VssProjectInfo;
                                     if (projectInfo == null || projectInfo.ContainsFiles())
                                     {
-                                        CaseSensitiveRename(sourcePath, targetPath, git.Move);
-                                        needCommit = true;
+                                        if (!dryRun)
+                                        {
+                                          CaseSensitiveRename(sourcePath, targetPath, git.Move);
+                                          needCommit = true;
+                                        }
                                     }
                                     else
                                     {
@@ -427,13 +452,19 @@ namespace Hpdi.Vss2Git
                                 {
                                     if (projectInfo.ContainsFiles())
                                     {
-                                        git.Move(sourcePath, targetPath);
-                                        needCommit = true;
+                                        if (!dryRun)
+                                        {
+                                          git.Move(sourcePath, targetPath);
+                                          needCommit = true;
+                                        }
                                     }
                                     else
                                     {
-                                        // git doesn't care about directories with no files
-                                        Directory.Move(sourcePath, targetPath);
+                                        if (!dryRun)
+                                        {
+                                            // git doesn't care about directories with no files
+                                            Directory.Move(sourcePath, targetPath);
+                                        }
                                     }
                                 }
                                 else
@@ -455,8 +486,11 @@ namespace Hpdi.Vss2Git
                                 project, target, moveToAction.NewProject);
                             if (projectInfo.Destroyed && targetPath != null && Directory.Exists(targetPath))
                             {
-                                // project was moved to a now-destroyed project; remove empty directory
-                                Directory.Delete(targetPath, true);
+                                if (!dryRun)
+                                {
+                                    // project was moved to a now-destroyed project; remove empty directory
+                                    Directory.Delete(targetPath, true);
+                                }
                             }
                         }
                         break;
@@ -520,7 +554,10 @@ namespace Hpdi.Vss2Git
                         }
                         else if (target.IsProject)
                         {
-                            Directory.CreateDirectory(targetPath);
+                            if (!dryRun)
+                            {
+                                Directory.CreateDirectory(targetPath);
+                            }
                             writeProject = true;
                         }
                         else
@@ -536,7 +573,10 @@ namespace Hpdi.Vss2Git
                         {
                             logger.WriteLine("{0}: Creating subdirectory {1}",
                                 projectDesc, projectInfo.LogicalName);
-                            Directory.CreateDirectory(projectInfo.GetPath());
+                            if (!dryRun)
+                            {
+                                Directory.CreateDirectory(projectInfo.GetPath());
+                            }
                         }
 
                         // write current rev of all contained files
@@ -557,8 +597,11 @@ namespace Hpdi.Vss2Git
                         if (WriteRevisionTo(target.PhysicalName, version, targetPath))
                         {
                             // add file explicitly, so it is visible to subsequent git operations
-                            git.Add(targetPath);
-                            needCommit = true;
+                            if (!dryRun)
+                            {
+                              git.Add(targetPath);
+                              needCommit = true;
+                            }
                         }
                     }
                 }
@@ -679,8 +722,11 @@ namespace Hpdi.Vss2Git
                 if (WriteRevisionTo(physicalName, version, path))
                 {
                     // add file explicitly, so it is visible to subsequent git operations
-                    git.Add(path);
-                    needCommit = true;
+                    if (!dryRun)
+                    {
+                        git.Add(path);
+                        needCommit = true;
+                    }
                 }
             }
             return needCommit;
@@ -690,12 +736,15 @@ namespace Hpdi.Vss2Git
         {
             VssFile item;
             VssFileRevision revision;
-            Stream contents;
+            Stream contents = null;
             try
             {
                 item = (VssFile)database.GetItemPhysical(physical);
                 revision = item.GetRevision(version);
-                contents = revision.GetContents();
+                if (!dryRun)
+                {
+                    contents = revision.GetContents();
+                }
             }
             catch (Exception e)
             {
@@ -706,10 +755,13 @@ namespace Hpdi.Vss2Git
                 return false;
             }
 
-            // propagate exceptions here (e.g. disk full) to abort/retry/ignore
-            using (contents)
+            if (!dryRun && null != contents)
             {
-                WriteStream(contents, destPath);
+                // propagate exceptions here (e.g. disk full) to abort/retry/ignore
+                using (contents)
+                {
+                    WriteStream(contents, destPath);
+                }
             }
 
             // try to use the first revision (for this branch) as the create time,
@@ -723,9 +775,15 @@ namespace Hpdi.Vss2Git
                 }
             }
 
-            // set file creation and update timestamps
-            File.SetCreationTimeUtc(destPath, TimeZoneInfo.ConvertTimeToUtc(createDateTime));
-            File.SetLastWriteTimeUtc(destPath, TimeZoneInfo.ConvertTimeToUtc(revision.DateTime));
+            DateTime createDateTimeUtc = TimeZoneInfo.ConvertTimeToUtc(createDateTime);
+            DateTime revisionDateTimeUtc = TimeZoneInfo.ConvertTimeToUtc(revision.DateTime);
+
+            if (!dryRun)
+            {
+                // set file creation and update timestamps
+                File.SetCreationTimeUtc(destPath, createDateTimeUtc);
+                File.SetLastWriteTimeUtc(destPath, revisionDateTimeUtc);
+            }
 
             return true;
         }
